+++ uptrace/uptrace.go
@@ -3,19 +3,12 @@ package uptrace
 import (
 	"context"
 	"os"
-	"runtime"
 	"sync/atomic"
-	"time"
 
 	"github.com/uptrace/uptrace-go/internal"
-	"github.com/uptrace/uptrace-go/metricexp"
-	"github.com/uptrace/uptrace-go/spanexp"
 
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/exporters/stdout/stdouttrace"
 	"go.opentelemetry.io/otel/propagation"
-	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -46,9 +39,9 @@ func ConfigureOpentelemetry(opts ...Option) {
 
 	client := newClient(dsn)
 
+	configurePropagator(cfg)
 	if !cfg.TracingDisabled {
-		configurePropagator(cfg)
-		configureTracing(client, cfg)
+		configureTracing(ctx, client, cfg)
 	}
 	if !cfg.MetricsDisabled {
 		configureMetrics(ctx, client, cfg)
@@ -66,75 +59,6 @@ func configurePropagator(cfg *config) {
 		)
 	}
 	otel.SetTextMapPropagator(textMapPropagator)
-}
-
-func configureTracing(client *client, cfg *config) {
-	client.provider = cfg.TracerProvider
-	if client.provider == nil {
-		var opts []sdktrace.TracerProviderOption
-
-		if res := cfg.newResource(); res != nil {
-			opts = append(opts, sdktrace.WithResource(res))
-		}
-		if cfg.TraceSampler != nil {
-			opts = append(opts, sdktrace.WithSampler(cfg.TraceSampler))
-		}
-
-		client.provider = sdktrace.NewTracerProvider(opts...)
-		otel.SetTracerProvider(client.provider)
-	}
-
-	spe, err := spanexp.NewExporter(&spanexp.Config{
-		DSN:            cfg.DSN,
-		Sampler:        cfg.TraceSampler,
-		BeforeSendSpan: cfg.BeforeSendSpan,
-	})
-	if err != nil {
-		internal.Logger.Printf("spanexp.NewExporter failed: %s", err)
-		return
-	}
-
-	queueSize := queueSize()
-	bsp := sdktrace.NewBatchSpanProcessor(spe,
-		sdktrace.WithMaxQueueSize(queueSize),
-		sdktrace.WithMaxExportBatchSize(queueSize),
-		sdktrace.WithBatchTimeout(5*time.Second),
-	)
-	client.provider.RegisterSpanProcessor(bsp)
-
-	if cfg.PrettyPrint {
-		exporter, err := stdouttrace.New(stdouttrace.WithPrettyPrint())
-		if err != nil {
-			internal.Logger.Printf(err.Error())
-		} else {
-			client.provider.RegisterSpanProcessor(sdktrace.NewSimpleSpanProcessor(exporter))
-		}
-	}
-}
-
-func configureMetrics(ctx context.Context, client *client, cfg *config) {
-	ctrl, err := metricexp.InstallNewPipeline(ctx, &metricexp.Config{
-		DSN: cfg.DSN,
-	}, controller.WithResource(cfg.newResource()))
-	if err != nil {
-		internal.Logger.Printf("metricexp.InstallNewPipeline failed: %s", err)
-		return
-	}
-	client.ctrl = ctrl
-}
-
-func queueSize() int {
-	const min = 1e3
-	const max = 10e3
-
-	n := runtime.GOMAXPROCS(0) * 1e3
-	if n < min {
-		return min
-	}
-	if n > max {
-		return max
-	}
-	return n
 }
 
 //------------------------------------------------------------------------------